pyproject.toml
[tool.poetry]
name = "gbb-practice"
version = "0.1.0"
description = "가위바위보 플레이 사진을 입력받아 승패를 판정하는 모델"
authors = ["rollforward"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
torch = { version = "^2.1.1+cu118", source="torch"}
torchvision = { version = "^0.16.1+cu118", source = "torch"}
torchaudio = { version = "^2.1.1+cu118", source = "torch"}
opencv-python = "^4.8.1.78"
matplotlib = "^3.8.2"
pyyaml = "^6.0.1"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
